switch to mitogen for ansible connection
the "mitogen" connection plugin uses python remote procedure calls
and can strongly speed up provisioning of VMs in remote regions.

in my tests, an example task (adding the "max" user to two new user
groups) took takes 15s on an AWS VM in the hongkong region, and 156s on
a Huawei VM in the Guangzhou region (and that's after turning on
"pipelining" to reduce the number of SSH connections).

switching to mitogen brought down the timings to 1s
for the AWS hongkong machine and 3s on the Huawei VM (i.e. more than one
order of magnitude).